# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi BKPSDM Katingan",
        "description": "Ini adalah dokumentasi API yang digunakan untuk website BKPSDM Katingan.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/unduhans": {
            "get": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Menampilkan daftar semua file unduhan",
                "operationId": "getUnduhanList",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Dokumen Penting"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Deskripsi singkat dokumen."
                                            },
                                            "file_path": {
                                                "type": "string",
                                                "readOnly": true
                                            },
                                            "original_filename": {
                                                "type": "string",
                                                "readOnly": true,
                                                "example": "nama_asli_dokumen.pdf"
                                            },
                                            "file_url": {
                                                "type": "string",
                                                "readOnly": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Mengunggah file baru (memerlukan token)",
                "operationId": "storeUnduhan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Judul File Baru"
                                    },
                                    "description": {
                                        "description": "Deskripsi file (opsional)",
                                        "type": "string",
                                        "example": "Ini adalah deskripsi file."
                                    },
                                    "file": {
                                        "description": "File yang akan diunggah (Maks 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File berhasil diunggah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Dokumen Penting"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Deskripsi singkat dokumen."
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "original_filename": {
                                            "type": "string",
                                            "readOnly": true,
                                            "example": "nama_asli_dokumen.pdf"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unduhans/{id}": {
            "get": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Menampilkan detail satu file unduhan",
                "operationId": "getUnduhanById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Dokumen Penting"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Deskripsi singkat dokumen."
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "original_filename": {
                                            "type": "string",
                                            "readOnly": true,
                                            "example": "nama_asli_dokumen.pdf"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Memperbarui data unduhan (tanpa file)",
                "operationId": "updateUnduhan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Judul File Diperbarui"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi diperbarui."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Dokumen Penting"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Deskripsi singkat dokumen."
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "original_filename": {
                                            "type": "string",
                                            "readOnly": true,
                                            "example": "nama_asli_dokumen.pdf"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Menghapus file (memerlukan token)",
                "operationId": "deleteUnduhan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "File berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unduhans/{id}/update-file": {
            "post": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Memperbarui file unduhan",
                "operationId": "updateUnduhanFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Judul File Diperbarui"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi diperbarui."
                                    },
                                    "file": {
                                        "description": "File baru untuk menggantikan yang lama (maksimal 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Dokumen Penting"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Deskripsi singkat dokumen."
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "original_filename": {
                                            "type": "string",
                                            "readOnly": true,
                                            "example": "nama_asli_dokumen.pdf"
                                        },
                                        "file_url": {
                                            "type": "string",
                                            "readOnly": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unduhans/{id}/download": {
            "get": {
                "tags": [
                    "Unduhan"
                ],
                "summary": "Mengunduh file unduhan",
                "operationId": "downloadUnduhanFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File berhasil diunduh",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "Data atau file tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint /api/login. Contoh: 1|aBcDeFgHiJkLmNoPqRsTuVwXyZ",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}